Require the Cell ID to be passed in
We noticed in nteract/nteract that we tend to want to know the ID after insertion
which either means that we pass the cell ID in or we change the return type for some functions.
This opts to pass the cell ID after.

Either way, we're going to have to release a major version since we're changing the call signatures.

Before we :ship: it though, we'll want to bring in some other utility functions that were written in
nteract and bring them to commutable.